c to P works
